+++ cognite/client/_api/data_modeling/instances.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import inspect
+import itertools
 import logging
 import random
 import time
@@ -48,6 +49,7 @@ from cognite.client.data_classes.data_modeling.instances import (
     EdgeApplyResultList,
     EdgeList,
     InstanceAggregationResultList,
+    InstanceInspectResultList,
     InstancesApplyResult,
     InstancesDeleteResult,
     InstanceSort,
@@ -711,6 +713,54 @@ class InstancesAPI(APIClient):
         node_ids = [NodeId.load(item) for item in deleted_instances if item["instanceType"] == "node"]
         edge_ids = [EdgeId.load(item) for item in deleted_instances if item["instanceType"] == "edge"]
         return InstancesDeleteResult(node_ids, edge_ids)
+
+    def inspect(
+        self,
+        nodes: NodeId | Sequence[NodeId] | tuple[str, str] | Sequence[tuple[str, str]] | None = None,
+        edges: EdgeId | Sequence[EdgeId] | tuple[str, str] | Sequence[tuple[str, str]] | None = None,
+        *,
+        return_all_view_versions: bool = False,
+        return_all_container_versions: bool = False,
+    ) -> InstanceInspectResultList:
+        """`Reverse lookup for instances. <https://developer.cognite.com/api/v1/#tag/Instances/operation/instanceInspect>`_
+
+        This method will return the involved views and containers for the given nodes and edges.
+
+        Args:
+            nodes (NodeId | Sequence[NodeId] | tuple[str, str] | Sequence[tuple[str, str]] | None): Node IDs.
+            edges (EdgeId | Sequence[EdgeId] | tuple[str, str] | Sequence[tuple[str, str]] | None): Edge IDs.
+            return_all_view_versions (bool): Include all view versions in inspection results.
+            return_all_container_versions (bool): Include all container versions in inspection results.
+
+        Returns:
+            InstanceInspectResultList: List of instance inspection results.
+
+        Examples:
+
+            Look up the involved views and containers for a given node and edge:
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes.data_modeling import NodeId, EdgeId
+                >>> client = CogniteClient()
+                >>> res = client.data_modeling.instances.inspect(
+                ...     nodes=NodeId("my-space", "foo1"),
+                ...     edges=EdgeId("my-space", "bar2"),
+                ... )
+        """
+        identifiers = self._load_node_and_edge_ids(nodes, edges)
+        options = {
+            "inspectionOperations": {
+                "involvedViews": {"allVersions": return_all_view_versions},
+                "involvedContainers": {"allVersions": return_all_container_versions},
+            }
+        }
+        # TODO: API claims limit is 1k, but only returns 200 max (<=100 for edges and similar for nodes). Thus
+        #       for now we chunk to just 100 elements:
+        items = itertools.chain.from_iterable(
+            self._post(self._RESOURCE_PATH + "/inspect", json={"items": chunk.as_dicts(), **options}).json()["items"]
+            for chunk in identifiers.chunked(100)
+        )
+        return InstanceInspectResultList._load(list(items), cognite_client=self._cognite_client)
 
     def subscribe(
         self,